@@ -1,8 +1,8 @@
 module github.com/sabitm/go-pretty
 
-go 1.17
+go 1.21
 
 require (
 	github.com/kr/text v0.2.0
-	github.com/rogpeppe/go-internal v1.9.0
+	github.com/rogpeppe/go-internal v1.13.1
 )
