@@ -2,14 +2,16 @@ package pretty_test
 
 import (
 	"fmt"
+
 	"github.com/sabitm/go-pretty"
 )
 
 func Example() {
+	//nolint:structcheck
 	type myType struct {
 		a, b int
 	}
-	var x = []myType{{1, 2}, {3, 4}, {5, 6}}
+	x := []myType{{1, 2}, {3, 4}, {5, 6}}
 	fmt.Printf("%# v", pretty.Formatter(x))
 	// output:
 	// []pretty_test.myType{
